fix-bug-cancel-order

+++ src/services/OrderService.js
@@ -128,24 +128,53 @@ const getOrderDetails = (id) => {
     })
 }
 
-const cancelOrderDetails = (id) => {
+const cancelOrderDetails = (id, data) => {
     return new Promise(async (resolve, reject) => {
         try {
-            const order = await Order.findByIdAndDelete(id)
-            if (order === null) {
+            let order = []
+            const promises = data.map(async (order) => {
+                const productData = await Product.findOneAndUpdate(
+                    {
+                    _id: order.product,
+                    selled: {$gte: order.amount}
+                    },
+                    {$inc: {
+                        countInStock: +order.amount,
+                        selled: -order.amount
+                    }},
+                    {new: true}
+                )
+                console.log('productData', productData)
+                if(productData) {
+                    order = await Order.findByIdAndDelete(id)
+                    if (order === null) {
+                        resolve({
+                            status: 'ERR',
+                            message: 'The order is not defined'
+                        })
+                    }
+                } else {
+                    return{
+                        status: 'OK',
+                        message: 'ERR',
+                        id: order.product
+                    }
+                }
+            })
+            const results = await Promise.all(promises)
+            const newData = results && results.filter((item) => item)
+            if(newData.length) {
                 resolve({
                     status: 'ERR',
-                    message: 'The order is not defined'
+                    message: `San pham voi id${newData.join(',')} khong ton tai`
                 })
             }
-
             resolve({
                 status: 'OK',
-                message: 'SUCESSS',
+                message: 'success',
                 data: order
             })
         } catch (e) {
-            console.log('e', e)
             reject(e)
         }
     })
